backend/requirements.txt
Django==5.2.9
djangorestframework==3.16.1
django-cors-headers==4.9.0
django-filter==25.2
celery==5.6.0
redis==7.1.0
django-celery-beat==2.8.1
requests==2.32.5
python-decouple==3.8
psycopg2-binary==2.9.9
gunicorn==22.0.0
whitenoise==6.7.0
django-environ==0.11.2
twilio==9.0.4
python-dateutil==2.8.2
